⭐ only show 8 lines in file.context by default (#5187)
First time using this with a large config I got a few screens of scrollback from trying to print the entire block of file.context. Most file context handlers however only ever want a few core content lines. We are starting with 8 by default, not sure yet if this is adequate or not.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>