⭐ epoch support in version (#5106)
* ⭐ epoch support in version

Add support for epochs in MQL's `semver` type.

```mql
> semver("1:1.2.3") < semver("2:1.0.0")
true
```

This works for both deb/rpm-style epochs ("NN:version") and python-style
epochs ([PEP440](https://peps.python.org/pep-0440/)).

Additionally you can access the epoch:

```mql
> semver("5:1.2.3").epoch
5

> semver("3!2.1").epoch
3
```

You can compare versions with and without epochs as well:

```mql
> semver("1:1.0") > semver("2009.12.03")
true
```

Note: Technically epochs aren't semver, so the name `semver` is hitting
its limit here. We could introduce the `version` type as a replacement
or as a temporary holdover (because we also don't want to clutter the
global namespace too much).

Internal note: This required a new wrapper around the version, which can
now be extended with more metadata. Currently we have the epoch stored.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

* 🧹 soft-deprecate semver in favor of version

The soft part of this deprecation is that we don't show a deprecation
warning anywhere at all. So semver is still technically included, since
we released it in v9.x, but we will deprecate it formally with the next
major release.

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>