🐛 fix core `asset.labels` + ✨  new `asset.platformMetadata` (#5227)
The labels of an asset come from the asset itself, not the platform, but as some point
we made a mistake to use the platform labels, this PR is the first step to fix this issue.

In this PR we are:

* Introducing a new field to the proto message `asset.Platform.Metadata` with the goal
to give more clarity and avoid confusions with `asset.Labels`
* We are deprecating `asset.Platform.Labels` in favor of `asset.Platform.Metadata` (this field
will be remove in `v12`)
* We are introducing a new `asset.platformMetadata` to the `asset` MQL schema, this
new field will have the `asset.Platform.Metadata`
* We are fixing the `asset.labels` MQL resource to have the actual asset labels but, we are
going to merge this field with the platform labels for backwards compatibility

After merging this PR, we will start migrating the use of `asset.Platform.Labels` to
`asset.Platform.Metadata`, so that in `v12` we can remove `asset.Platform.Labels.

---------

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>