⭐ ip.inRange (#5271)
After: https://github.com/mondoohq/cnquery/pull/5265/

This adds a new function to the IP type to check if it is in range of a subnet or two IP addresses. 

This is how to use it with IPv4 subnets:

```coffee
> ip('192.2.3.4').inRange('192.2.3.0/24')
true

# We can also auto-detect the subnet if you don't specify it. For example, this is a class C address, so:
> ip('192.2.3.4').inRange('192.2.3.0')
true

> ip('192.2.3.4').inRange('192.1.1.0/24')
false
```

Here is how you use it with 2 IP addresses:

```coffee
> ip('192.2.3.4').inRange('192.2.3.1', '192.2.3.5')
true

> ip('192.2.3.4').inRange('192.2.3.5', '192.2.3.255')
false

> ip('2001:db8:3c4d:15::1a2f:1a2b').inRange('2001:db8:3c4d::', '2001:db8:3c4e::')
true
```

This PR also improves the auto-detection of netmasks. In IPv4 we can now auto-detect class A, B, and C networks and is able to print the subnet mask via the `subnet` field:

```coffee
> ip('1.2.3.4').subnet
"255.0.0.0"

> ip('172.0.0.1').subnet
"255.255.0.0"

> ip('192.168.0.1').subnet
"255.255.255.0"
```

For IPv6 the `subnet` works differently. From my current understanding it is the subnet-ID that most people care about. The concept of subnet masks doesn't exist here [[1](https://en.wikipedia.org/wiki/IP_address#Subnetworks)]. The subnet is the part of the 64-bit network prefix that is not used by the routing prefix. For example:

```coffee
# 64-bit, default IPv6 prefix:
> ip("2001:db8:3c4d:15::1a2f:1a2b").prefixLength
64
> ip("2001:db8:3c4d:15::1a2f:1a2b").prefix
"2001:db8:3c4d:15::"
> ip("2001:db8:3c4d:15::1a2f:1a2b").subnet
""

# 48-bit prefix:
> ip("2001:db8:3c4d:15::1a2f:1a2b/48").prefixLength
48
> ip("2001:db8:3c4d:15::1a2f:1a2b/48").prefix
"2001:db8:3c4d::"
> ip("2001:db8:3c4d:15::1a2f:1a2b/48").subnet
"15"
```

As you can see, the `prefix` remains in full IPv6 notation, even if the lower bits are appreviated with `"::"`. The subnet, however, is only the identifier since it is not commonly represented as a full IPv6 address, unlike the prefix and host.

Other small fixes:
- compare IP, string and dict types (e.g. `ip("1.2.3.4") == "1.2.3.4"`)
- support the subnet mask of `"/0"` (it was trying to detect the default subnet before, even if users specified that they explicitly wanted this subnet)