⭐️ os cloud resource for azure (#5305)
Part 3 of https://github.com/mondoohq/cnquery/pull/5284

This PR enables the new os resource `cloud` in Azure VMs.

Additional changes:

* ✨ update Azure IMDS API version to `2023-07-01`

We are not using version 2023-11-15 since it is still being rolled out and it may not
be available in some regions.

Docs:

https://learn.microsoft.com/en-us/azure/virtual-machines/instance-metadata-service?tabs=windows#supported-api-versions

* ⚙️ make broadcast and gateway ip types
* 🧹 pretty cloud providers name
* ✨ do not fail when no `loadbalancer/` data

The loadbalancer/ endpoint from the IMDS can come back with the
following output:
```
{ "error": "No load balancer metadata is found. Please check if your VM is using any non-basic SKU load balancer and retry later." }
```

In these cases, we won't error the `RawMetadata()` function call since
we already fetched the `instance/` metadata endpoint.

---------

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>