🐛 mqlc: compile builtin functions without constraint block
Fixes https://github.com/mondoohq/cnquery/issues/5248

Tracing this issue from the beginning of the compilation until the end,
I noticed that our compiler does the right thing by not passing any
constraint block when there is none, the issue is in the post compilation
where we expand any builtin function block  `"$one", "$all", "$none", "$any"`.

This is a very small change that checks if there is a valid block to
expand, and if not, we continue the processing of other blocks.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>