✨ Safe manipulation of `inventory.Platform` (#5417)
One of the core data structures of Mondoo’s plugin system is the `inventory.Asset` struct that contains details about an asset, including platform information that is stored in `Asset.Platform`.

The `inventory.Platform` struct stores important information used by the Mondoo platform (`server`) like the runtime, the platform kind, and additional metadata.

When we run `cnquery` or `cnspec` most of the times the first thing we do is an initial discovery to detect the asset and platform we are going to scan, then we pass these core data structures to our providers where, sometimes, we do additional discovery and enrichment of data. An example of this is the discovery of AWS instances (`aws` provider) that later gets enriched with additional information from the `os` provider.

**Change**

We are introducing two accessors. One at the Asset level and the other one directly into the Platform struct. These new accessor should be used through out the code base to safely manipulate the platform of an asset.

---------

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>