✨ cache aws connection and upstream client
This is an attempt to reduce the load of the AWS provider. We are going
to cache the upstream client which uses ranger under the hood and
creates goroutines to fulfill the requests. This is not a problem for a
few runtimes, but when we do thousands of them, we really don't need to
create that many clients.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>