🧪 experiment: avoid caching `aws` clients
The caching of clients for the `aws` provider assumes that we will
cache clients globally but in reality we cache clients per runtime,
and when we have thousands of runtimes we are essentially storing
thousands of clients in memory.

Before removing or reimplementing this code, we will try to disable the
caching and measure.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>