⭐️ `lr`: add provider dependencies to schema (#5495)
* ⭐️ `lr`: add provider dependencies to schema

Resources within a provider can depend on resources from another
provider, for example; The `os` provider depends on `network` provider.

The `lr` parser already reads these imports but it wasn't passing it
to the schema. This change adds the list of dependencies so that we can
install them at runtime.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>

* ✨ test dependencies in lr files

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>

---------

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>