⭐️ add provider dependencies as builtin (#5503)
This change uses the new schema changes added at #5495 which exposes
dependencies that need to be consider in two scenarios:

1. Developing providers
2. Air-gapped environments

Now, when configuring a builtin provider, we will look at its
dependencies and add them to the providers configuration and build them
too.

Remote dependencies are not yet supported, we need to revisit this
change once we support them.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>