🐛 detect if tls is enabled (#5531)
The `tls` resource is part of multiple core providers, it requires a `socket`
to know the host, port and protocol, many fields of this resource require
that the server (configured in the socket) speaks TLS and not plain text
(like HTTP), therefore this change makes it so that we only detect once if
the socket is running on TLS or not. If not, we return null properly to
indicate that there are no certificates, if it does, then we collect TLS
information.

Additionally, I am setting a default timeout of one minute to dial to the
socket compared to what was before (unlimited=no timeout).

Finally, if the TLS tester returns a timeout, we don't error the detection
but instead, we set the fields as null.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>