🐛 support inRange for nil values in maps
This was previously producing errors or would outright crash with a panic if the value was not set. Maps have fixed types, so it looks for typed functions to do the actual test. However, since a key of a map may not exist, the typed value may still be nil.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>