🐛 detect OpenSSH 9.9 and force Key Exchange Algorithms (#5568)
OpenSSH 9.9 introduces hybrid post-quantum algorithms, and
when users try to connect via SSH, they get a key signature
compatibility issue, most likely due to the go package
https://pkg.go.dev/golang.org/x/crypto/ssh not supporting it yet
or misinterpreting it.

Note that, there is also a reported bug that recommends
upgrading to OpenSSH 9.9p2

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1103392

We recommend users to upgrade their servers to that version.

Regardless, this change introduces an extra step during an SSH
connection that detects if the server's advertised key exchange
methods include unsupported hybrid ones, and if so, we will
force the key exchange algorithm to a compatible one.

---------

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>