⭐️ enable cloud resource for EBS volume scanning (#5554)
**💁‍♂️ Problem/Context**

We recently shipped cloud and cloud IP detection in Mondoo for agent-based scans on cloud hosts. This feature has been very useful in properly presenting hosts as running in the cloud and has also enabled us to expose internet connected hosts via a new risk factor.

Unfortunately if the user is scanning their infrastructure using the EBS snapshot scanning capabilities in the AWS Serverless integration, they will not get this data back. While we cannot fetch the metadata endpoint during the EBS volume scan we can populate much of the cloud resource data including:

* Asset.kind is virtualmachine
* Provider (obviously this is going to be AWS)
* Local IP / Public IP (via further inspection of the AWS instance via API calls)

**📖 User Story**

As a user I want to understand the configuration data of all the assets in my business, so that I can better respond to Mondoo findings.

---------

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>