🐛 fix long-standing print issues in assertions (#5683)
* 🐛 fix long-standing print issues on assertions

This is a start to tackle some long-standing printing issues we have experienced with results.

Take the following example:

```sh
> cnquery run -c "terraform.blocks.none(arguments.source == /google/)" terraform tf
```

The current output will look like this:

```coffee
[failed] [].none()
  actual:   [
    0: terraform.block type="module" labels=[
  0: terraform.block "project-factory"
] {
      arguments: {
        source: "terraform-google-modules/project-factory/google"
        version: "14.4.0"
      }
      arguments.source: "terraform-google-modules/project-factory/google"
    }
  ]
```

As you can see, the printing of the `actual` value is really broken. If
we focus on the problem, we see it's related to how the labels are
printed.

Here is the fixed version:

```coffee
[failed] [].none()
  actual:   [
    0: terraform.block type="module" labels=["project-factory"] {
      arguments: {
        source: "terraform-google-modules/project-factory/google"
        version: "14.4.0"
      }
      arguments.source: "terraform-google-modules/project-factory/google"
    }
  ]
```

In the fix I realized that both the array-printer as well as ref-printer
are both broken. They were both referencing the first entrypoint of the
first block of code in every code bundle. Of course the printer is
cycling through all data and thus this is incorrect as soon as we start
to print anything but the first entrypoint.

The other issue was that codeIDs (in actuality checksums) were
incorrectly propagated. The code bundle indeed contains a codeID, but we
can't use them, because auto-expansion uses checksums. The new code for
expanding list assertions also uses checksums now, so both are aligned.

This isn't the end of the issue, it's just the beginning. The immediate
follow-up will address the unnecessary resolution of the comparable, as
seen in the example posted above (printing of `arguments`). This gets
pretty bad with terraform particulary.

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>