✨ introduce human time (#5778)
This allows users to specify timestamps in a human-readable way. Particularly useful when dealing with YAML files, ie querypacks and policies, where you can now specify timestamps with a lot more grace and variety.

We are also breaking out the time parsing code into its own utility so the same parser is now shared between MQL and policy/querypack bundles.

Finally I found an issue with the numeric variants of RFC 822 and RFC 1123, since both are parsed in their non-numeric variants without error. Thus, the numeric variants are removed for now in the auto-detection.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>