🔨 overhaul featureflags: configuration + activation + metadata
This is a full overhaul of featureflags. They have now moved to features.yaml and can be configured with the new featureflags utility. This utility makes sure we never mix up IDs for features, sets up default and available features, and generates some helper go-code around featureflags.

The reason we are adding this now is:
- the use of default featureflags was inconsistent between different code paths
- the use of feature-loading from CLI and env variables was inconsistent as well
- a series of features that should have been set as default features by now have not been set - they were simply overlooked
- the list of features still contains a lot of names that have been builtin or sunset now i.e. setting them has no effect at all -> we can now identify what is usable and what is not

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>